rename usdc.n to usdc
