--- deployments/arbitrum/usdc/roots.json
@@ -1,1 +1,0 @@
-{}
